+++ tests/testthat/test-chrono.subsets.R
@@ -133,15 +133,6 @@ time_subsets <- chrono.subsets.continuous(data, tree, time, model = "deltran", F
 ## Test
 test_that("chrono.subsets.continuous works properly with deltran model", {
     
-    # Get slice
-    # expect_equal(
-    #     as.vector(na.omit(unlist(get.slice(1, time[2], "ACCTRAN", adjust.FADLAD(FADLAD, tree, data), data, verbose = FALSE))))
-    #     , c(7, 8, 9, 1, 2, 3, 12, 13, 14, 15, 44, 70, 73, 76, 79, 85, 48, 90, 47, 95, 46, 98))
-    # expect_equal(
-    #     unname(unlist(get.slice(slice = 1, time = time[2], model = "proximity", ages_tree = adjust.FADLAD(FADLAD, tree, data), data = data, verbose = FALSE, tree = tree)))
-    #     , c(7, 8, 9, 1, 2, NA, 12, 13, 14, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
-    # )
-
     ## class is list
     expect_is(
         time_subsets
@@ -152,17 +143,20 @@ test_that("chrono.subsets.continuous works properly with deltran model", {
         , 3)
     ## elements per subsets
     subsets_1 <- c("n1","n5","n11", "Zhangolestes")
+    # c(11), sort(c(51, 55, 61))
     subsets_2 <- c("n6","Asioryctes","n7","n8","n10","n15","n17","n19","n22","n25","n34","n39","n44","n47")
+    # c(8), sort(c(56,57,58,60,65,67,69,72,75,84,89,94,97))
     subsets_3 <- c("n17","n23","n27","n28","n29","n31","n32","n39","n42","n44","n48","n49")
+    # sort(c(67,73,77,78,79,81,82,89,92,94,98,99))
     expect_equal(
-        rownames(data[time_subsets[[1]]$elements,])
-        , subsets_1)
+        sort(rownames(data[time_subsets[[1]]$elements,]))
+        , sort(subsets_1))
     expect_equal(
-        rownames(data[time_subsets[[2]]$elements,])
-        , subsets_2)
+        sort(rownames(data[time_subsets[[2]]$elements,]))
+        , sort(subsets_2))
     expect_equal(
-        rownames(data[time_subsets[[3]]$elements,])
-        , subsets_3)
+        sort(rownames(data[time_subsets[[3]]$elements,]))
+        , sort(subsets_3))
 })
 
 ## ACCTRAN
@@ -174,14 +168,14 @@ test_that("chrono.subsets.continuous works properly with acctran model", {
     subsets_2 <- c("Kennalestes","Asioryctes","Ukhaatherium","Cimolestes","Maelestes","Batodon","Zalambdalestes","Barunlestes","Gypsonictops","Leptictis","Oxyclaenus","n20","n23","n26","n29","n35","Cynocephalus","n40","Patriomanis","n45","Icaronycteris","n48")
     subsets_3 <- c("Leptictis","Dasypodidae","n24","Potamogalinae","Dilambdogale","Widanelfarasia","Rhynchocyon","Procavia","Moeritherium","Trichechus","Cynocephalus","Adapis","Patriomanis","Soricidae","Solenodon")
     expect_equal(
-        rownames(data[time_subsets[[1]]$elements,])
-        , subsets_1)
+        sort(rownames(data[time_subsets[[1]]$elements,]))
+        , sort(subsets_1))
     expect_equal(
-        rownames(data[time_subsets[[2]]$elements,])
-        , subsets_2)
+        sort(rownames(data[time_subsets[[2]]$elements,]))
+        , sort(subsets_2))
     expect_equal(
-        rownames(data[time_subsets[[3]]$elements,])
-        , subsets_3)
+        sort(rownames(data[time_subsets[[3]]$elements,]))
+        , sort(subsets_3))
 })
 
 ## chrono.subsets
@@ -316,7 +310,6 @@ test_that("Sanitizing works for chrono.subsets (wrapper)", {
     rownames(data_wrong)[1] <- "wrong!"
     error <- capture_error(chrono.subsets(data = data_wrong, tree = BeckLee_tree, method = "c", time = 3, model = "acctran"))
     expect_equal(error[[1]], "The labels in the matrix and in the tree do not match!\nTry using clean.data() to match both tree and data or make sure whether nodes should be included or not (inc.nodes = FALSE by default).")
-
 })
 
 test_that("Output format is correct", {
@@ -472,20 +465,20 @@ test_that("probability models work", {
     data(BeckLee_ages)
     data(BeckLee_tree)
     
-    test1 <- chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = c(120, 100, 80, 60, 40 , 20, 0), model = "gradual.split", inc.nodes = TRUE, BeckLee_ages, verbose = FALSE, t0 = FALSE)
+    test1 <- chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = c(120, 100, 80, 60, 40 , 20, 0), model = "gradual.split", inc.nodes = TRUE, FADLAD = BeckLee_ages, verbose = FALSE, t0 = FALSE)
     test2 <- chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = c(120, 100, 80, 60, 40 , 20, 0), model = "equal.split", inc.nodes = TRUE, BeckLee_ages, verbose = FALSE, t0 = FALSE)
 
     expect_is(test1, "dispRity")
     expect_is(test1$subsets[[1]][[1]], "matrix")
     expect_equal(dim(test1$subsets[[1]][[1]]), c(6,3))
     expect_true(all(test1$subsets[[1]][[1]][,1:2] >= 1))
-    expect_true(all(test1$subsets[[1]][[1]][,3] < 1))
+    expect_true(all(test1$subsets[[1]][[1]][,3] <= 1))
 
     expect_is(test2, "dispRity")
     expect_is(test2$subsets[[1]][[1]], "matrix")
     expect_equal(dim(test2$subsets[[1]][[1]]), c(6,3))
     expect_true(all(test2$subsets[[1]][[1]][,1:2] >= 1))
-    expect_true(all(test2$subsets[[1]][[1]][,3] < 1))
+    expect_true(all(test2$subsets[[1]][[1]][,3] <= 1))
 })
 
 test_that("chrono.subsets detects distance matrices", {
@@ -569,7 +562,7 @@ test_that("chrono.subsets works with multiPhylo", {
     expect_equal(names(test$subsets), c("9.31405078347417", "4.65702539173708", "0"))
     expect_equal(unique(unlist(lapply(test$subsets, names), use.names = FALSE)), "elements")
     expect_equal(unlist(lapply(test$subsets, lapply, dim), use.names = FALSE), c(3, 2, 5, 2, 10, 2))
-    expect_equal(unique(c(test$subsets[[2]]$elements)), c(17, 22, 26, 21, NA, 2, 25, 27))
+    expect_equal(unique(c(test$subsets[[2]]$elements)), c(17, 22, 21, 26, NA, 2, 25, 27))
 
     ## Works with discrete
     test <- chrono.subsets(data, tree, method = "discrete", time = 3, inc.nodes = TRUE)
@@ -596,42 +589,9 @@ test_that("chrono.subsets works with multiple matrices", {
     expect_equal(length(test$matrix), 2)
     expect_equal(dim(test$matrix[[1]]), c(50, 48))
 
-
-    set.seed(1)
-    ## Matches the trees and the matrices
-    ## A bunch of trees
-    make.tree <- function(n, fun = rtree) {
-        ## Make the tree
-        tree <- fun(n)
-        tree <- chronos(tree, quiet = TRUE,
-                        calibration = makeChronosCalib(tree, age.min = 10, age.max = 10))
-        class(tree) <- "phylo"
-        ## Add the node labels
-        tree$node.label <- paste0("n", 1:Nnode(tree))
-        ## Add the root time
-        tree$root.time <- max(tree.age(tree)$ages)
-        return(tree)
-    }
-    trees <- replicate(3, make.tree(10), simplify = FALSE)
-    class(trees) <- "multiPhylo"
-
-
-    ## A bunch of matrices
-    ## Base matrix
-    matrix_base <- matrix(rnorm(30), 10, 3, dimnames = list(paste0("t", 1:10)))
-    do.ace <- function(tree, matrix) {
-        ## Run one ace
-        fun.ace <- function(character, tree) {
-            results <- ace(character, phy = tree)$ace
-            names(results) <- paste0("n", 1:Nnode(tree))
-            return(results)
-        }
-        ## Run all ace
-        return(rbind(matrix, apply(matrix, 2, fun.ace, tree = tree)))
-    }
-
-    ## All matrices
-    matrices <- lapply(trees, do.ace, matrix_base)
+    load("bound_test_data.Rda")
+    trees <- bound_test_data$trees
+    matrices <- bound_test_data$matrices
 
     ## Test if it works with multiple trees and with multiple matrices ok
     test <- chrono.subsets(matrices, tree = trees[[1]], time = 3, method = "continuous", model = "acctran", t0 = 5)
@@ -664,14 +624,7 @@ test_that("chrono.subsets works with multiple matrices", {
     expect_is(test$subsets, "list")
     expect_equal(length(test$subsets), 3)
     expect_equal(dim(test$subsets$`5`$elements), c(7, 3))
-
-
-    ## Calculating disparity works
-    
-
 })
-
-
 
 test_that("fast internal functions work", {
 
@@ -683,11 +636,13 @@ test_that("fast internal functions work", {
     colnames(edge_table) <- c("parent", "child")
     tree$tip.label <- paste0("t", 1:Ntip(tree))
     tree$node.label <- paste0("n", (Ntip(tree)+1):(Ntip(tree)+Nnode(tree)))
+    data <- matrix(NA, ncol = 1, nrow = 11)
+    rownames(data) <- c(tree$tip.label, tree$node.label)
     # plot(tree)
     # nodelabels(tree$node.label)
     # edgelabels(rownames(edge_table))
     # axisPhylo()
-    # slice <- 2
+    # slice <- 3
     # abline(v = tree$root.time - slice)
     # edge_table
 
@@ -713,7 +668,7 @@ test_that("fast internal functions work", {
     expect_equal(test[1,], c(7, 0.25, 8, 0.75))
     expect_equal(test[2,], c(7, 0.25, 6, 2.75))
 
-    ## Tip slices
+    ## Tip slices
     t4 <- test <- fast.slice.table(0, tree)
     expect_is(test, "matrix")
     expect_equal(dim(test), c(4,4))
@@ -788,76 +743,46 @@ test_that("fast internal functions work", {
     expect_equal(select.table.tips(t5, "gradual.split"), cbind(t5[,c(1,3)], 1-(t5[,2]/(t5[,2]+t5[,4]))))
     expect_equal(select.table.tips(t6, "gradual.split"), cbind(t6[,c(1,3)], 1-(t6[,2]/(t6[,2]+t6[,4]))))
 
-    # ## get.time.slice works
-    # test1 <- get.time.slice(time = 3.5, tree, model = "deltran")
-    # expect_is(test1, "list")
-    # expect_equal(names(test1), "elements")
-    # expect_equal(test1[[1]], matrix(c(9, 8 ,7)))
-    # test2 <- get.time.slice(time = 3.5, tree, model = "gradual.split")
-    # expect_is(test2, "list")
-    # expect_equal(names(test2), "elements")
-    # expect_equal(test2[[1]], cbind(t1[,c(1,3)], 1-(t1[,2]/(t1[,2]+t1[,4]))))    
+    ## get.time.slice works
+    test1 <- get.time.slice(time = 3.5, tree, model = "deltran", verbose = FALSE)
+    expect_is(test1, "list")
+    expect_equal(names(test1), "elements")
+    expect_equal(test1[[1]], matrix(c(9, 8 ,7)))
+    test2 <- get.time.slice(time = 3.5, tree, model = "gradual.split", verbose = FALSE)
+    expect_is(test2, "list")
+    expect_equal(names(test2), "elements")
+    expect_equal(test2[[1]], cbind(t1[,c(1,3)], 1-(t1[,2]/(t1[,2]+t1[,4]))))
 
-    # ## add.FADLAD works
-    # FADLAD <- matrix(c(3, 1.5, 2, 0, 5, 4), 3, 2, byrow = TRUE, dimnames = list(c("t4", "t3", "t6"), c("FAD", "LAD")))
-    # test1 <- get.time.slice(time = 2, tree, model = "proximity")
-    # test2 <- get.time.slice(time = 2, tree, model = "equal.split")
-
-    # res1 <- add.FADLAD(test1, 2, FADLAD, tree)
-    # expect_is(res1, "list")
-    # expect_equal(names(res1), "elements")
-    # expect_equal(res1[[1]], matrix(c(10, 5 , 3, 4)))
-
-    # res2 <- add.FADLAD(test2, 2, FADLAD, tree)
-    # expect_is(res2, "list")
-    # expect_equal(names(res2), "elements")
-    # expect_equal(res2[[1]][,1], c(10,10,8,3,4))
-    # expect_equal(res2[[1]][,3], c(.5,.5,.5,1,1))
+    ## add.FADLAD works
+    FADLAD <- matrix(c(3, 1.5, 2, 0, 5, 4), 3, 2, byrow = TRUE, dimnames = list(c("t4", "t3", "t6"), c("FAD", "LAD")))
+    test1 <- get.time.slice(time = 2, tree, model = "proximity", verbose = FALSE)
+    test2 <- get.time.slice(time = 2, tree, model = "equal.split", verbose = FALSE)
 
 
-#     ## Speed benchmarking
-#     library(microbenchmark)
-#     microbenchmark(
-#         chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = 5, model = "gradual.split", inc.nodes = TRUE, FADLAD = BeckLee_ages, verbose = FALSE, algo.fast = TRUE),
-#         chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = 5, model = "gradual.split", inc.nodes = TRUE, FADLAD = BeckLee_ages, verbose = FALSE, algo.fast = FALSE)
-#     )
+    res1 <- add.FADLAD(test1, 2, FADLAD, rownames(data))
+    expect_is(res1, "list")
+    expect_equal(names(res1), "elements")
+    expect_equal(res1[[1]], matrix(c(10, 5 , 3, 4)))
 
-# test_fast <- chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = 5, model = "gradual.split", inc.nodes = TRUE, FADLAD = BeckLee_ages, verbose = FALSE, algo.fast = TRUE)
-# test_slow <- chrono.subsets(BeckLee_mat99, BeckLee_tree, method = "continuous", time = 5, model = "gradual.split", inc.nodes = TRUE, FADLAD = BeckLee_ages, verbose = FALSE, algo.fast = FALSE)
+    res2 <- add.FADLAD(test2, 2, FADLAD, rownames(data))
+    expect_is(res2, "list")
+    expect_equal(names(res2), "elements")
+    expect_equal(res2[[1]][,1], c(10,3,8,4))
+    expect_equal(res2[[1]][,3], c(.5,1,.5,1))
 
-# data = BeckLee_ages
-# tree = BeckLee_tree
-# method = "continuous"
-# time = 5
-# model = "gradual.split"
-# inc.nodes = FALSE
-# FADLAD = BeckLee_ages
-# verbose = FALSE
-# t0 = FALSE
-# bind.data = FALSE
-# algo.fast = TRUE
+    ## FADLAD works with nodes
+    FADLAD <- rbind(FADLAD, "n10" = c(3, 1))
+    res1 <- add.FADLAD(test1, 2, FADLAD, rownames(data))
+    expect_is(res1, "list")
+    expect_equal(names(res1), "elements")
+    expect_equal(res1[[1]], matrix(c(10, 5 , 3, 4)))
 
-## PROBLEM WITH *.SPLIT remove duplicated selected species when probablility of sampling == 1 (e.g. when FADLAD) - test with BeckLee data slice 0.
-
-
-# expect_equal(test_fast$subset[[1]], test_slow$subset[[1]])
-# expect_equal(test_fast$subset[[2]], test_slow$subset[[2]])
-# c(tree$tip.label, tree$node.label)[(test_slow$subset[[2]][[1]])]
-# c(tree$tip.label, tree$node.label)[(test_fast$subset[[2]][[1]])] # OK
-
-# expect_equal(test_fast$subset[[3]], test_slow$subset[[3]])
-# c(tree$tip.label, tree$node.label)[(test_slow$subset[[3]][[1]])]
-# c(tree$tip.label, tree$node.label)[(test_fast$subset[[3]][[1]])] # OK
-
-# expect_equal(test_fast$subset[[4]], test_slow$subset[[4]])
-# c(tree$tip.label, tree$node.label)[(test_slow$subset[[4]][[1]])]
-# c(tree$tip.label, tree$node.label)[(test_fast$subset[[4]][[1]])]
-
-# expect_equal(sort(test_fast$subset[[5]][[1]]), sort(test_slow$subset[[5]][[1]]))
-# c(tree$tip.label, tree$node.label)[(test_slow$subset[[4]][[1]])]
-# c(tree$tip.label, tree$node.label)[(test_fast$subset[[4]][[1]])]
-
-
+    res2 <- add.FADLAD(test2, 2, FADLAD, rownames(data))
+    expect_is(res2, "list")
+    expect_equal(names(res2), "elements")
+    expect_equal(res2[[1]][,1], c(10,3,8,4))
+    expect_equal(res2[[1]][,2], c(10,3,5,4))
+    expect_equal(res2[[1]][,3], c(1,1,.5,1))
 })